Removed keepass, and put creds in a dot file for no. Fixed some other bugs in installer.

+++ build_grid.sh
@@ -19,15 +19,19 @@ local_iso="iso/9front.$iso_arch.iso"
 [ ! -d img ] && mkdir img
 [ ! -d bin ] && mkdir bin
 
-###########
-# Keepass #
-###########
-# Create keepass db(belagos.kdbx) so we dont have to default passwords.
-[ ! -f belagos.kdbx ] && build_grid/keepass.exp
+###################
+# Glenda Password #
+###################
 
-# Get it again for use below
-keepass_passphrase=''
-read -p "Enter password for keepass(belagos.kdbx) again[echoed]: " keepass_passphrase
+if [ ! -f .belagos_pass ]; then
+   belagos_pass_default=`< /dev/urandom tr -dc _A-Z-a-z-0-9 | head -c10`
+   echo "Passord will be stored in .belagos_pass"
+   read -p "Enter password for glenda(default $belagos_pass_default): " belagos_pass
+   [ -z $belagos_pass ] && belagos_pass=$belagos_pass_default
+   touch .belagos_pass
+   chmod 600 .belagos_pass
+   echo $belagos_pass > .belagos_pass
+fi
 
 ################
 # Download ISO #
@@ -52,32 +56,30 @@ echo "qemu-system-$qemu_arch $kvm -m 256 -net nic,macaddr=52:54:00:00:EE:03 -net
 chmod u+x bin/run_fsserve.sh
 
 # This can be large, as all other VMs will boot off of this disk
-#qemu-img create -f qcow2 img/9front_fsserve.img 10G
-
-# Creating base install image
-#build_grid/9front_install.exp bin/run_fsserve.sh $local_iso
+if [ ! -f img/9front_fsserve.img ]; then
+   qemu-img create -f qcow2 img/9front_fsserve.img 10G
+   # Creating base install image
+   build_grid/9front_install.exp bin/run_fsserve.sh $local_iso
+fi
 
 # Back it up as expect is unrelyable with curses
-cp img/9front_fsserve.img img/9front_fsserve.img_back
+[ ! -f img/9front_fsserve.img_back ] && cp img/9front_fsserve.img img/9front_fsserve.img_back
 
 # Set up networking and turn on PXE
-echo "$keepass_passphrase" | build_grid/9front_fsserve.exp bin/run_fsserve.sh
+cat .belagos_pass | build_grid/9front_fsserve.exp bin/run_fsserve.sh
 
 # Boot fsserve in the background and wait until its up
 bin/run_headless.exp bin/run_fsserve.sh > /dev/null 2>&1 &
 sleep 10
-echo "$keepass_passphrase" | build_grid/9front_boot_wait.sh 192.168.9.3
+bin/9front_boot_wait.sh 192.168.9.3
 
-# Run installer
-echo "$keepass_passphrase" | build_grid/9front_fsserve_net_and_pxe.exp
-
-#echo "$keepass_passphrase" | build_grid/keepass_get.exp glenda | /opt/drawterm/drawterm -h 192.168.9.3 -a 192.168.9.3 -u glenda -G -c /mnt/term/$PWD/rc/fsserve_build.rc
-#echo "$keepass_passphrase" | build_grid/keepass_get.exp glenda | /opt/drawterm/drawterm -h 192.168.9.3 -a 192.168.9.3 -u glenda -G -c ip/tftpd
+# Run installer via drawterm
+cat .belagos_pass | build_grid/9front_fsserve_net_and_pxe.exp
 
 # Boot fsserve in the background and wait until its up
 bin/run_headless.exp bin/run_fsserve.sh > /dev/null 2>&1 &
 sleep 10
-echo "$keepass_passphrase" | build_grid/9front_boot_wait.sh 192.168.9.3
+bin/9front_boot_wait.sh 192.168.9.3
 
 #############
 # AUTHSERVE #
@@ -88,12 +90,12 @@ echo "qemu-system-$qemu_arch $kvm -m 256 -net nic,macaddr=52:54:00:00:EE:04 -net
 chmod u+x bin/run_authserve.sh
 
 qemu-img create -f qcow2 img/9front_authserve.img 1M
-echo "$keepass_passphrase" | build_grid/9front_authserve.exp bin/run_authserve.sh
+cat .belagos_pass | build_grid/9front_authserve.exp bin/run_authserve.sh
 
 # Run it in the BG as we need it for cpuserve creation
 bin/run_headless.exp bin/run_authserve.sh > /dev/null 2>&1 &
 sleep 10
-echo "$keepass_passphrase" | build_grid/9front_boot_wait.sh 192.168.9.4
+bin/9front_boot_wait.sh 192.168.9.4
 
 ############
 # CPUSERVE #
@@ -103,11 +105,11 @@ echo "qemu-system-$qemu_arch $kvm -smp 4 -m 256 -net nic,macaddr=52:54:00:00:EE:
 chmod u+x bin/run_cpuserve.sh
 
 qemu-img create -f qcow2 img/9front_cpuserve.img 1M
-echo "$keepass_passphrase" | build_grid/9front_authserve.exp bin/run_cpuserve.sh
+cat .belagos_pass | build_grid/9front_cpuserve.exp bin/run_cpuserve.sh
 
 ########################
 # Turn down everything #
 ########################
 
-echo "$keepass_passphrase" | build_grid/keepass_get.exp glenda | /opt/drawterm/drawterm -h 192.168.9.4 -a 192.168.9.4 -u glenda -G -c "fshalt"
-echo "$keepass_passphrase" | build_grid/keepass_get.exp glenda | /opt/drawterm/drawterm -h 192.168.9.3 -a 192.168.9.3 -u glenda -G -c "fshalt"
+cat .belagos_pass | /opt/drawterm/drawterm -h 192.168.9.4 -a 192.168.9.4 -u glenda -G -c "fshalt"
+cat .belagos_pass | /opt/drawterm/drawterm -h 192.168.9.3 -a 192.168.9.3 -u glenda -G -c "fshalt"
